# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  version: 0.1.0
  title: Adwa AI Tour Guide — Admin API
  description: "The full admin surface: auth, museums, admin users, rooms, items,
    audit, overview, billing, and the visitor ticket check. Generated from Zod
    schemas — do not hand-edit. Every JSON request body is capped at 100kb
    globally; routes that accept a body therefore declare 413. See
    docs/backend-implementation-plan.md and developer1-detailed-plan.md."
servers:
  - url: /
    description: Relative to wherever the API is deployed
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin JWT from POST /admin/login. See §8.
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
              example: Invalid credentials.
            code:
              $ref: "#/components/schemas/ErrorCode"
            requestId:
              type: string
              example: 01HQZXK7N8VJYB3F9G6R2T4W5M
            details:
              type: array
              items:
                type: object
                properties:
                  path:
                    type: string
                  message:
                    type: string
                required:
                  - path
                  - message
              description: Present only for VALIDATION_ERROR, one entry per invalid field.
          required:
            - message
            - code
            - requestId
      required:
        - error
      description: Shape of every non-2xx response, from every route. See §7.1.
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNAUTHENTICATED
        - INVALID_CREDENTIALS
        - FORBIDDEN
        - CROSS_TENANT_ACCESS
        - NOT_FOUND
        - CONFLICT
        - ROOM_REFERENCED
        - INVALID_ROOM_SEQUENCE
        - RATE_LIMITED
        - INTERNAL_ERROR
        - UPSTREAM_FAILURE
        - UPSTREAM_UNAVAILABLE
        - TIER_LIMIT_EXCEEDED
        - SUBSCRIPTION_INACTIVE
        - PAYMENT_ALREADY_PENDING
        - PAYMENT_NOT_FOUND
        - TICKET_URL_INVALID
      description: Machine-readable error code. See §7.2 of the backend implementation plan.
...
    AdminUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type:
            - string
            - "null"
        role:
          type: string
          enum:
            - SYSTEM_ADMIN
            - MUSEUM_ADMIN
        status:
          type: string
          enum:
            - ACTIVE
            - INVITED
            - SUSPENDED
        museumId:
          type:
            - string
            - "null"
          format: uuid
        museumName:
          type:
            - string
            - "null"
        lastLoginAt:
          type:
            - string
            - "null"
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - displayName
        - role
        - status
        - museumId
        - museumName
        - lastLoginAt
        - createdAt
...
    ListAdminsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminUser"
        nextCursor:
          type:
            - string
            - "null"
      required:
        - data
        - nextCursor
    AdminStatus:
      type: string
      enum:
        - ACTIVE
        - INVITED
        - SUSPENDED
    CreateAdminRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        displayName:
          type: string
          minLength: 1
          maxLength: 200
        role:
          type: string
          enum:
            - SYSTEM_ADMIN
            - MUSEUM_ADMIN
        museumId:
          type: string
          format: uuid
          description: Required when role is MUSEUM_ADMIN.
      required:
        - email
        - password
        - role
    UpdateAdminRequest:
      type: object
      properties:
        displayName:
          type:
            - string
            - "null"
          minLength: 1
          maxLength: 200
        status:
          $ref: "#/components/schemas/AdminStatus"
        password:
          type: string
          minLength: 8
      additionalProperties: false
...
paths:
...
  /admin/admins:
    get:
      tags:
        - Admins
      summary: List admin accounts
      description: A MUSEUM_ADMIN sees only its own museum’s seats and never an
        operator account; scope comes from the token, not from ?museumId.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
            description: Page size. Defaults to 50, capped at 200.
          required: false
          description: Page size. Defaults to 50, capped at 200.
          name: limit
          in: query
        - schema:
            type: string
            description: Opaque cursor from a previous page.
          required: false
          description: Opaque cursor from a previous page.
          name: cursor
          in: query
        - schema:
            type: string
            format: uuid
            description: SYSTEM_ADMIN only. A MUSEUM_ADMIN is always scoped to its own
              museum.
          required: false
          description: SYSTEM_ADMIN only. A MUSEUM_ADMIN is always scoped to its own museum.
          name: museumId
          in: query
        - schema:
            type: string
            enum:
              - SYSTEM_ADMIN
              - MUSEUM_ADMIN
          required: false
          name: role
          in: query
        - schema:
            $ref: "#/components/schemas/AdminStatus"
          required: false
          name: status
          in: query
        - schema:
            type: string
            minLength: 1
            maxLength: 200
          required: false
          name: search
          in: query
      responses:
        "200":
          description: Paginated list of admin accounts.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAdminsResponse"
        "400":
          description: Malformed or missing request fields (VALIDATION_ERROR).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Missing, invalid, or expired token (UNAUTHENTICATED /
            INVALID_CREDENTIALS).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      tags:
        - Admins
      summary: Create an admin account (SYSTEM_ADMIN only)
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAdminRequest"
      responses:
        "201":
          description: Account created, in INVITED status until its first login.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          description: Malformed or missing request fields (VALIDATION_ERROR).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Missing, invalid, or expired token (UNAUTHENTICATED /
            INVALID_CREDENTIALS).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Wrong role or wrong museum scope (FORBIDDEN / CROSS_TENANT_ACCESS).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Resource does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Unique constraint conflict, or a room delete blocked by a reference
            (CONFLICT / ROOM_REFERENCED).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "413":
          description: Request body exceeds the 100kb cap (VALIDATION_ERROR).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /admin/admins/{id}:
    patch:
      tags:
        - Admins
      summary: Update an admin’s display name, status, or password
      description: "Role is deliberately not editable: moving a seat across the tenant
        boundary in one PATCH would be a silent privilege change. Delete and
        re-create instead."
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAdminRequest"
      responses:
        "200":
          description: Updated account.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "400":
          description: Malformed or missing request fields (VALIDATION_ERROR).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Missing, invalid, or expired token (UNAUTHENTICATED /
            INVALID_CREDENTIALS).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Wrong role or wrong museum scope (FORBIDDEN / CROSS_TENANT_ACCESS).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Resource does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Unique constraint conflict, or a room delete blocked by a reference
            (CONFLICT / ROOM_REFERENCED).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "413":
          description: Request body exceeds the 100kb cap (VALIDATION_ERROR).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      tags:
        - Admins
      summary: Delete an admin account
      description: Refused for your own account, and for a museum’s last active
        administrator — both would strand someone.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            format: uuid
          required: true
          name: id
          in: path
      responses:
        "204":
          description: Account deleted. Its audit trail is retained.
        "401":
          description: Missing, invalid, or expired token (UNAUTHENTICATED /
            INVALID_CREDENTIALS).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Wrong role or wrong museum scope (FORBIDDEN / CROSS_TENANT_ACCESS).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Resource does not exist.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Unique constraint conflict, or a room delete blocked by a reference
            (CONFLICT / ROOM_REFERENCED).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...